Added header and footer, updated files
Updated wiki and addArticle with the header and footer. Updated wiki.php to have a sidebar and updated the connection code to run because it was broken. Updated style.css to account for the sidebar. Separated the header and footer from index.php so they could be used on all pages.